Llama2Inference

Documentation for Llama2Inference.

Llama2Inference.get_most_common_pair

Llama2Inference.replace_top_pair!

Llama2Inference.rmsnorm!

Llama2Inference.softmax!

Llama2Inference.get_most_common_pair — Method

get_most_common_pair(vocab_ids::Vector{Int})

Get the most common consecutive pair of vocabulary IDs in the training data.

Arguments

vocab_ids::Vector{Int}: Vector of the trainings text's vocabulary IDs.

Returns

top_pair::Tuple{Int, Int}, max_count::Int: Tuple containing the most common pair and its count.

Calls

count_consecutive : Gets the dictionary containing number of occurrences of each consecutive pairs.

Examples

text = "hello" tokenizer = Tokenizer(text) toppair, maxcount = getmostcommonpair(tokenizer.vocabids)

Llama2Inference.replace_top_pair! — Method

replace_top_pair!(vocab_ids::Vector{Int}, vocab::OrderedDict{Int, Vector{UInt8}})

Replaces the 20 most common consecutive pairs with a new vocabulary ID.

Arguments

vocab_ids::Vector{Int}: Vector of the trainings text's vocabulary IDs.

vocab::OrderedDict{Int, Vector{UInt8}}: Ordered dictionary mapping IDs to vocabulary bytes.

Calls

get_most_common_pair : Gets the most common pair.

Examples

text = "hello" tokenizer = Tokenizer(text) replacetoppair!(tokenizer.vocab_ids, tokenizer.vocab)

Llama2Inference.rmsnorm! — Method

rmsnorm!(out::Array{T, 1}, x::Array{T,1}, weight::Array{T,1}) where T<:AbstractFloat

normalize out in place by the root mean square of x and multiply by the learned weights weight.

Llama2Inference.softmax! — Method

softmax!(x::T{T2,1}) where {T<:AbstractArray, T2<:AbstractFloat}

softmax the values in x in place, up to position pos inclusively.