Llama2Inference

Documentation for Llama2Inference.

Llama2Inference.Sampler

Llama2Inference.forward

Llama2Inference.generate

Llama2Inference.rmsnorm!

Llama2Inference.sample

Llama2Inference.sample_topp

Llama2Inference.softmax!

Llama2Inference.forward — Method

forward(transformer::Transformer, state::RunState, token::Int, pos::Int)

Forward the token at position pos through the transformer model.

The foward pass corresponds to the LlaMa2 decoder architecture and is based on the C implementation by Andrej Karpathy.

Arguments

transformer::Transformer: The transformer object containg config, weights and the token embedding table.

state::RunState: The state object to store the intermediate results during the forward pass.

token::Int: The token to forward through the transformer.

pos::Int: The position of the token in the sequence. The position is 1-based, which means the first position in a sequence is 1.

Example

julia> config, weights = read_checkpoint("./stories15M.bin")
julia> transformer = Transformer(config, weights)
julia> tokenizer = build_tokenizer("./tokenizer.bin", Int(config.vocab_size))
julia> state = RunState(config)
julia> token = 2
julia> pos = 1
julia> forward(transformer, state, token, pos)
32000-element Vector{Float32}:
 -6.790799
  0.828116
 -6.790441
 -6.7904925
 -6.7904897
 -6.7906227
  ⋮
 -6.7905755
 -6.7905526
 -6.7907033
 -6.790716
 -6.7906933
 -6.790564

Llama2Inference.generate — Method

generate(transformer::Transformer, tokenizer::Tokenizer, sampler::Sampler, steps::Int; prompt::String="")

Generate a sequence of tokens using the transformer.

Arguments

transformer::Transformer: The transformer object containg config and weights.

tokenizer::Tokenizer: The tokenizer object to encode and decode tokens.

sampler::Sampler: The sampler object to sample a token from the output logits.

steps::Int: The number of maximum tokens to generate.

prompt::String: The input text to start the generation. If none, the generation starts with an empty string.

Example

julia> config, weights = read_checkpoint("./stories15M.bin")
julia> transformer = Transformer(config, weights)
julia> tokenizer = build_tokenizer("./tokenizer.bin", Int(config.vocab_size))
julia> sampler = Sampler(config.vocab_size, 0.0f0, 0.9f0)
julia> generate(transformer, tokenizer, sampler, 23; prompt="The universe")
The universe was bright and full of stars. Every night, the stars would twinkle and shine.

Llama2Inference.rmsnorm! — Method

rmsnorm!(out::AbstractArray{Float32, 1}, x::AbstractArray{Float32,1}, weight::AbstractArray{Float32,1})

Normalize out in place by the root mean square of x and multiply with weight.

\[out_i = \frac{x_i}{RMS(x)} * weight_i \quad\text{,where}\quad RMS(x)= \sqrt{ (\frac{1}{n} * \sum_{i=1}^{n} x_i^2) +1\mathrm{e}{-5}} \]

1e-5 is added for numerical stability in the square root part.

Llama2Inference.softmax! — Method

softmax!(x::AbstractArray{Float32,1})

Softmax the values in x in place.

\[x_i = \frac{e^{x_i}}{\sum_{j=1}^{n} e^{x_j}}\]

Llama2Inference.Sampler — Method

Sampler(vocab_size::Int, temperature::Float32, topp::Float32)

A struct to hold the parameters for sampling from a distribution.

Arguments

vocab_size::Int: The size of the vocabulary.

temperature::Float32: The temperature to apply to the logits before sampling.

topp::Float32: The probability mass to sample from the top-p distribution.

Llama2Inference.sample — Method

sample(sampler::Sampler, logits::Vector{Float32})

Sample from the logits using the sampler.

Example

julia> sampler = Sampler(6, 0.0f0, 0.0f0)
julia> logits = [0.1f0, 0.3f0, 0.2f0, 0.15f0, 0.15f0, 0.1f0]
julia> sample(sampler, logits)
2

Llama2Inference.sample_topp — Method

sample_topp(sampler::Sampler, logits::Vector{Float32}, coin::Float32)::Int

top-p sampling (or "nucleus sampling") samples from the smallest set of tokens that exceed probability topp.