Llama2Inference

Documentation for Llama2Inference.

Llama2Inference.forward

Llama2Inference.rmsnorm!

Llama2Inference.sample_topp

Llama2Inference.softmax!

Llama2Inference.forward — Method

forward(transformer::Transformer, state::RunState, token::Int, pos::Int)

Forward pass of the transformer model with the input token at position pos.

LlaMa2 was used as the architecure of the transformer model and it's modifications. The forward pass looks like the following:

forward through all layers

a) RMSNorm
b) linear project x into Query, Key, Value with Wq, Wk, Wv
c) RoPE relative positional encoding
d) multihead attention
e) residual of x + RMSNorm
f) MLP with SwiGLU non-linearity

rmsnrom

classify into logits

Llama2Inference.rmsnorm! — Method

rmsnorm!(out::AbstractArray{Float32, 1}, x::AbstractArray{Float32,1}, weight::AbstractArray{Float32,1})

Normalize out in place by the root mean square of x and multiply with weight. 1e-5 is added for numerical stability in the square root part.

\[out_i = rac{x_i}{RMS(x)} * weight_i, where RMS(x) = \sqrt{ rac{1}{n} * \sum_{i=1}^{n} x_i^2} + 1e-5}\]

Llama2Inference.sample_topp — Method

sample_topp(sampler::Sampler, logits::Vector{Float32}, coin::Float32)::Int

top-p sampling (or "nucleus sampling") samples from the smallest set of tokens that exceed probability topp.

Llama2Inference.softmax! — Method

softmax!(x::AbstractArray{Float32,1})

Softmax the values in x in place.

\[x_i = rac{e^{x_i}}{\sum_{j=1}^{n} e^{x_j}}\]