Llama2Inference

Documentation for Llama2Inference.

Llama2Inference.Config

Llama2Inference.RunState

Llama2Inference.Sampler

Llama2Inference.TokenIndex

Llama2Inference.Tokenizer

Llama2Inference.Transformer

Llama2Inference.TransformerWeights

Llama2Inference.build_tokenizer

Llama2Inference.decode

Llama2Inference.encode

Llama2Inference.find_token_id

Llama2Inference.find_token_str

Llama2Inference.forward!

Llama2Inference.generate

Llama2Inference.get_weights

Llama2Inference.is_ascii_string

Llama2Inference.mat_T_vec!

Llama2Inference.memory_map_weights

Llama2Inference.read_checkpoint

Llama2Inference.read_config

Llama2Inference.rmsnorm!

Llama2Inference.safe_print

Llama2Inference.sample

Llama2Inference.sample_topp

Llama2Inference.set_config_vocab_size

Llama2Inference.softmax!

Llama2Inference.sort_vocab!

Llama2Inference.Config — Type

Config{T<:Integer}

Struct to hold the configuration of the Transformer used for inference. It contains the following fields:

dim::T: transformer dimension

hidden_dim::T: for ffn layers

n_layers::T: number of layers

n_heads::T: number of query heads

nkvheads::T: number of key/value heads

vocab_size::T: vocabulary size

seq_len::T: max sequence length

It is possible to construct a Config using the constructor

Config{T}(dim::T, hidden_dim::T, n_layers::T, n_heads::T, n_kv_heads::T, vocab_size::T, seq_len::T)

but it is recommended to use the data format defined by Andrew Karpathy to store both the configuration and weights of the transformer used and read the config and weights using read_checkpoint.

Llama2Inference.RunState — Method

RunState(config::Config; T::Type{<:AbstractFloat} = Float32)

Create a RunState object with fields of type T (Float32 by default) from the given configuration.

This is used to store the state of the model during inference and is 0-initialized.

Llama2Inference.Sampler — Method

Sampler(vocab_size::T, temperature::Z, topp::Z) where {T<:Signed, Z<:AbstractFloat}

Create a sampler object with the given parameters, used for sampling from a distribution.

Arguments

vocab_size::T: The size of the vocabulary, will be casted to Int32.

temperature::Z: The temperature to apply to the logits before sampling, will be casted to Float32.

topp::Z: The probability mass to sample from the top-p distribution, will be casted to Float32.

Llama2Inference.TokenIndex — Type

struct TokenIndex

A data structure representing a token and its corresponding index.

Fields

str::String: The token as a string.

id::Int: The ID of the token.

Llama2Inference.Tokenizer — Type

struct Tokenizer

A data structure representing a tokenizer with its vocabulary and related properties.

Fields

vocab_size::Int: The size of the vocabulary.

vocab::Vector{String}: A vector containing the vocabulary tokens.

vocab_scores::Vector{Float32}: A vector containing scores associated with each token in the vocabulary.

max_token_length::Int: The maximum length of a token.

sorted_vocab::Union{Nothing, Vector{TokenIndex}}: Can be either nothing or a vector of TokenIndex structs representing the tokens and their respective indices.

Llama2Inference.Transformer — Type

Transformer struct which contains the fields

config::Config

weights::TransformerWeights

Llama2Inference.TransformerWeights — Type

TransformerWeights{T<:AbstractFloat}

Struct to hold the weights of the Transformer used for inference.

It is possible to construct a TransformerWeights using the constructor

TransformerWeights{T<:AbstractFloat}(
    token_embedding_table::AbstractArray{T,2}, 
    rms_att_weight::AbstractArray{T,2}, 
    wq::AbstractArray{T,3}, 
    wk::AbstractArray{T,3}, 
    wv::AbstractArray{T,3},
    wo::AbstractArray{T,3},
    rms_ffn_weight::AbstractArray{T,2},
    w1::AbstractArray{T,3},
    w2::AbstractArray{T,3},
    w3::AbstractArray{T,3},
    rms_final_weight::AbstractArray{T,1},
    wcls::AbstractArray{T,2}
    )

but it is recommended to use the data format defined by Andrew Karpathy to store both the configuration and weights of the transformer used and read the config and weights using read_checkpoint.

Llama2Inference.build_tokenizer — Method

build_tokenizer(filepath::String, vocab_size::Int)

Constructs a Tokenizer from a file.

Arguments

filepath::String: Path to the file containing vocabulary data.

vocab_size::Int: Size of the vocabulary to read from the file.

Returns

Tokenizer: A Tokenizer instance with the necessary data.

Summary Steps

Initialization

Reading file

Handling errors

Closing file

Returning Tokenizer with necessary data

Llama2Inference.decode — Method

decode(tokenizer::Tokenizer, prev_token::Int, token::Int) -> Union{String, UInt8}

Decodes a token ID into its corresponding token string representation or byte value using the Tokenizer.

Arguments

tokenizer::Tokenizer: The Tokenizer containing vocabulary and token mapping.

prev_token::Int: Token ID of the previous token in the sequence.

token::Int: Token ID to decode into its corresponding token string representation.

Returns

byte_val or token_str: Returns a decoded string or a byte value, depending on the type of the given token.

Description

Finds the token string using find_token_str.

Removes leading whitespace if the previous token is BOS.

Checks for raw byte tokens and parses them if applicable.

Returns the token string or its byte value representation.

Llama2Inference.encode — Method

encode(tokenizer::Tokenizer, text::String, use_bos::Bool, use_eos::Bool) -> Vector{Int}

Encodes the input text into a sequence of token IDs using the provided Tokenizer.

Arguments

tokenizer::Tokenizer: The Tokenizer containing vocabulary and token mapping.

text::String: Input text to encode into tokens.

use_bos::Bool: Indicates if a token ID representing the beginning of the sequence should be included.

use_eos::Bool: Indicates if a token ID representing the end of the sequence should be included.

Returns

Vector{Int}: Returns a vector of token IDs representing the encoded input text.

Description

Ensures the tokenizer's vocabulary is sorted with sort_vocab!.

Encodes text into bytes.

Initializes the token ID vector.

Optionally adds the BOS token.

Handles leading whitespace.

Looks up and stores token IDs with find_token_id.

Performs merges (BPE) based on scores (vocab_scores).

Optionally adds the EOS token.

Returns the vector tokens_indices representing the encoded input text.

Llama2Inference.find_token_id — Method

find_token_id(tokenizer::Tokenizer, token_str::String)

Finds and returns the ID of a given token string in the sorted_vocab.

Arguments

tokenizer::Tokenizer: The Tokenizer to search within.

token_str::String: The token string to search for its ID.

Return

Int: The ID of the token if found, otherwise returns -1.

Description

Iterates over the sorted_vocab to find the token string and return its ID.

Llama2Inference.find_token_str — Method

findtokenstr(tokenizer::Tokenizer, token_id::Int)

Finds and returns the token string coressponding to a given token ID from sort_vocab

Arguments

tokenizer::Tokenizer: Tokenizer to search within

token_id::Int: ID of the given token to find its coressponding string

Return

token_index.str: String of the token if found, otherwise returns nothing

Description

Iterates over the sorted_vocab to find the token ID and return its corresponding string.

Llama2Inference.forward! — Method

forward!(transformer::Transformer, state::RunState, token::Int, pos::Int)

Forward the token at position pos through the transformer model.

The foward pass corresponds to the LlaMa2 decoder architecture and is based on the C implementation by Andrej Karpathy. The output is a logits vector of size transformer.config.vocab_size.

Arguments

transformer::Transformer: The transformer object containg config, weights and the token embedding table.

state::RunState: The state object to store the intermediate results during the forward pass.

token::Int: The token to forward through the transformer.

pos::Int: The position of the token in the sequence. The position is 1-based, which means the first position in a sequence is 1.

Example

julia> config, weights = read_checkpoint("./bin/stories15M.bin")
julia> transformer = Transformer(config, weights)
julia> tokenizer = build_tokenizer("./bin/tokenizer.bin", Int(config.vocab_size))
julia> state = RunState(config)
julia> token = 2
julia> pos = 1
julia> forward!(transformer, state, token, pos)
julia> state.logits
32000-element Vector{Float32}:
 -6.7907834
  0.82811606
 -6.7904234
 -6.790472
  ⋮
 -6.79068
 -6.790696
 -6.7906737
 -6.7905493

Llama2Inference.generate — Method

generate(transformer::Transformer, tokenizer::Tokenizer, sampler::Sampler, steps::Int; prompt::String="", performance=true)

Generate a sequence of tokens using the transformer.

Arguments

transformer::Transformer: The transformer object containg config and weights.

tokenizer::Tokenizer: The tokenizer object to encode and decode tokens.

sampler::Sampler: The sampler object to sample a token from the output logits.

steps::Int: The number of maximum tokens to generate, upper bound by transformer.config.seq_len.

prompt::String: The input text to start the generation. If none, the generation starts with an empty string.

performance::Bool: If true, print the number of generated tokens and number of tokens generated per second.

Example

julia> config, weights = read_checkpoint("./bin/stories15M.bin")
julia> transformer = Transformer(config, weights)
julia> tokenizer = build_tokenizer("./bin/tokenizer.bin", Int(config.vocab_size))
julia> sampler = Sampler(config.vocab_size, 0.0f0, 0.9f0)
julia> generate(transformer, tokenizer, sampler, 23; prompt="The universe", performance=true)
The universe was bright and full of stars. Every night, the stars would twinkle and shine.

Llama2Inference.get_weights — Method

get_weights(weights::Vector, offset_1, offset_2, dims)::AbstractArray

Creates a view of weights from offset_1 to offset_2 with the dimension specified by dims.

Llama2Inference.is_ascii_string — Method

is_ascii_string(s::String) -> Bool

Check if all characters in the given string are ASCII characters.

Arguments

s::String: The input string to be checked.

Returns

Bool: Returns true if all characters in the string are ASCII characters, otherwise returns false.

Examples

Llama2Inference.mat_T_vec! — Method

mat_T_vec!(out::AbstractArray{T,1}, x::AbstractArray{T,1}, w::AbstractArray{T,2}) where T<:AbstractFloat

Efficient transpose(matrix)-vector multiplication out = w' * x.

Llama2Inference.memory_map_weights — Method

memory_map_weights(config::Config, weights::Vector{T}, shared_weights::Int32)::TransformerWeights{T} where {T<:AbstractFloat}

Takes the values in weights and maps creates a new TransformerWeights struct using config and shared_weights.

Llama2Inference.read_checkpoint — Function

read_checkpoint(checkpoint::String, T_weights::Type=Float32, T_config::Type=Int32)::Tuple{Config{T_config},TransformerWeights{T_weights}}

Reads the config and weights from the file at checkpoint.

Llama2Inference.read_config — Method

read_config(T::Type{<:Integer}, file::IO)::Config{T}

Reads from file to create a Config{T}.

Llama2Inference.rmsnorm! — Method

rmsnorm!(out::AbstractArray{Float32, 1}, x::AbstractArray{Float32,1}, weight::AbstractArray{Float32,1})

Normalize out in place by the root mean square of x and multiply with weight.

\[out_i = \frac{x_i}{RMS(x)} * weight_i \quad\text{,where}\quad RMS(x)= \sqrt{ (\frac{1}{n} * \sum_{i=1}^{n} x_i^2) +1\mathrm{e}{-5}} \]

1e-5 is added for numerical stability in the square root part.

Llama2Inference.safe_print — Method

safe_print(piece::String)

Prints the string piece to the console, skipping any "unsafe" bytes.

Llama2Inference.sample — Method

sample(sampler::Sampler, logits::Vector{Float32})

Sample from the logits using the sampler.

When temperature is 0, this function returns the index of the maximum value in logits.

When temperature is not 0 the following approach is used:

Apply the temperature to the logits.

Apply softmax to the logits to get a probability distribution.

IF 0 < topp < 1 THEN use top-p ("nucleus") sampling ELSE sample from the full distribution.

Example

julia> sampler = Sampler(6, 0.0, 0.0)
julia> logits = [0.1f0, 0.3f0, 0.2f0, 0.15f0, 0.15f0, 0.1f0]
julia> sample(sampler, logits)
2

Llama2Inference.sample_topp — Method

sample_topp(sampler::Sampler, logits::Vector{Float32}, coin::Float32)::Int

top-p sampling (or "nucleus sampling") samples from the smallest set of tokens that exceed probability topp.

Llama2Inference.set_config_vocab_size — Method

set_config_vocab_size(config::Config{T}, vocab_size::T)

Creates a new Config{T} with the field vocab_size updated to vocab_size.

Llama2Inference.softmax! — Method

softmax!(x::AbstractArray{Float32,1})

Softmax the values in x in place.

\[x_i = \frac{e^{x_i}}{\sum_{j=1}^{n} e^{x_j}}\]

Llama2Inference.sort_vocab! — Method

sort_vocab!(tokenizer::Tokenizer)

Sorts the vocabulary of the given Tokenizer, storing unique tokens and sorting them.

Argument

tokenizer::Tokenizer: The Tokenizer whose vocabulary is to be sorted.

Summary Steps

Checking if sorted_vocab is nothing or empty.

Initialization of data structures.

Identification of unique tokens.

Sorting the vocabulary.