Llama2Inference

Documentation for Llama2Inference.

Llama2Inference.Config

Llama2Inference.RunState

Llama2Inference.Sampler

Llama2Inference.Transformer

Llama2Inference.TransformerWeights

Llama2Inference.forward!

Llama2Inference.generate

Llama2Inference.get_weights

Llama2Inference.mat_T_vec!

Llama2Inference.memory_map_weights

Llama2Inference.read_checkpoint

Llama2Inference.read_config

Llama2Inference.rmsnorm!

Llama2Inference.sample

Llama2Inference.sample_topp

Llama2Inference.set_config_vocab_size

Llama2Inference.softmax!

Llama2Inference.Transformer — Type

Transformer struct which contains the fields

config::Config

weights::TransformerWeights

Llama2Inference.forward! — Method

forward!(transformer::Transformer, state::RunState, token::Int, pos::Int)

Forward the token at position pos through the transformer model.

The foward pass corresponds to the LlaMa2 decoder architecture and is based on the C implementation by Andrej Karpathy. The output is a logits vector of size transformer.config.vocab_size.

Arguments

transformer::Transformer: The transformer object containg config, weights and the token embedding table.

state::RunState: The state object to store the intermediate results during the forward pass.

token::Int: The token to forward through the transformer.

pos::Int: The position of the token in the sequence. The position is 1-based, which means the first position in a sequence is 1.

Example

julia> config, weights = read_checkpoint("./stories15M.bin")
julia> transformer = Transformer(config, weights)
julia> tokenizer = build_tokenizer("./tokenizer.bin", Int(config.vocab_size))
julia> state = RunState(config)
julia> token = 2
julia> pos = 1
julia> forward!(transformer, state, token, pos)
julia> state.logits
32000-element Vector{Float32}:
 -6.7907834
  0.82811606
 -6.7904234
 -6.790472
  ⋮
 -6.79068
 -6.790696
 -6.7906737
 -6.7905493

Llama2Inference.generate — Method

generate(transformer::Transformer, tokenizer::Tokenizer, sampler::Sampler, steps::Int; prompt::String="", performance=true)

Generate a sequence of tokens using the transformer.

Arguments

transformer::Transformer: The transformer object containg config and weights.

tokenizer::Tokenizer: The tokenizer object to encode and decode tokens.

sampler::Sampler: The sampler object to sample a token from the output logits.

steps::Int: The number of maximum tokens to generate, upper bound by transformer.config.seq_len.

prompt::String: The input text to start the generation. If none, the generation starts with an empty string.

performance::Bool: If true, print the number of generated tokens and number of tokens generated per second.

Example

julia> config, weights = read_checkpoint("./stories15M.bin")
julia> transformer = Transformer(config, weights)
julia> tokenizer = build_tokenizer("./tokenizer.bin", Int(config.vocab_size))
julia> sampler = Sampler(config.vocab_size, 0.0f0, 0.9f0)
julia> generate(transformer, tokenizer, sampler, 23; prompt="The universe", performance=true)
The universe was bright and full of stars. Every night, the stars would twinkle and shine.

Llama2Inference.mat_T_vec! — Method

mat_T_vec!(out::AbstractArray{T,1}, x::AbstractArray{T,1}, w::AbstractArray{T,2}) where T<:AbstractFloat

Efficient transpose(matrix)-vector multiplication out = w' * x.

Llama2Inference.read_checkpoint — Function

read_checkpoint(checkpoint::String, T_weights::Type=Float32, T_config::Type=Int32)::Tuple{Config{T_config},TransformerWeights{T_weights}}

Reads the config and weights from the file at checkpoint.

Llama2Inference.rmsnorm! — Method

rmsnorm!(out::AbstractArray{Float32, 1}, x::AbstractArray{Float32,1}, weight::AbstractArray{Float32,1})

Normalize out in place by the root mean square of x and multiply with weight.

\[out_i = \frac{x_i}{RMS(x)} * weight_i \quad\text{,where}\quad RMS(x)= \sqrt{ (\frac{1}{n} * \sum_{i=1}^{n} x_i^2) +1\mathrm{e}{-5}} \]

1e-5 is added for numerical stability in the square root part.

Llama2Inference.softmax! — Method

softmax!(x::AbstractArray{Float32,1})

Softmax the values in x in place.

\[x_i = \frac{e^{x_i}}{\sum_{j=1}^{n} e^{x_j}}\]

Llama2Inference.Sampler — Method

Sampler(vocab_size::Int, temperature::Float32, topp::Float32)

Create a sampler object with the given parameters, used for sampling from a distribution.

Arguments

vocab_size::Int: The size of the vocabulary.

temperature::Float32: The temperature to apply to the logits before sampling.

topp::Float32: The probability mass to sample from the top-p distribution.

Llama2Inference.sample — Method

sample(sampler::Sampler, logits::Vector{Float32})

Sample from the logits using the sampler.

Example

julia> sampler = Sampler(6, 0.0f0, 0.0f0)
julia> logits = [0.1f0, 0.3f0, 0.2f0, 0.15f0, 0.15f0, 0.1f0]
julia> sample(sampler, logits)
2

Llama2Inference.sample_topp — Method

sample_topp(sampler::Sampler, logits::Vector{Float32}, coin::Float32)::Int

top-p sampling (or "nucleus sampling") samples from the smallest set of tokens that exceed probability topp.

Llama2Inference.Config — Type

Config{T<:Integer}

Struct to hold the configuration of the Transformer used for inference. It contains the following fields:

dim::T: transformer dimension

hidden_dim::T: for ffn layers

n_layers::T: number of layers

n_heads::T: number of query heads

nkvheads::T: number of key/value heads

vocab_size::T: vocabulary size

seq_len::T: max sequence length

It is possible to construct a Config using the constructor

Config{T}(dim::T, hidden_dim::T, n_layers::T, n_heads::T, n_kv_heads::T, vocab_size::T, seq_len::T)

but it is recommended to use the data format defined by Andrew Karpathy to store both the configuration and weights of the transformer used and read the config and weights using read_checkpoint.

Llama2Inference.read_config — Method

read_config(T::Type{<:Integer}, file::IO)::Config{T}

Reads from file to create a Config{T}.

Llama2Inference.set_config_vocab_size — Method

set_config_vocab_size(config::Config{T}, vocab_size::T)

Creates a new Config{T} with the field vocab_size updated to vocab_size.

Llama2Inference.TransformerWeights — Type

TransformerWeights{T<:AbstractFloat}

Struct to hold the weights of the Transformer used for inference.

It is possible to construct a TransformerWeights using the constructor

TransformerWeights{T<:AbstractFloat}(
    token_embedding_table::AbstractArray{T,2}, 
    rms_att_weight::AbstractArray{T,2}, 
    wq::AbstractArray{T,3}, 
    wk::AbstractArray{T,3}, 
    wv::AbstractArray{T,3},
    wo::AbstractArray{T,3},
    rms_ffn_weight::AbstractArray{T,2},
    w1::AbstractArray{T,3},
    w2::AbstractArray{T,3},
    w3::AbstractArray{T,3},
    rms_final_weight::AbstractArray{T,1},
    wcls::AbstractArray{T,2}
    )

but it is recommended to use the data format defined by Andrew Karpathy to store both the configuration and weights of the transformer used and read the config and weights using read_checkpoint.

Llama2Inference.get_weights — Method

get_weights(weights::Vector, offset_1, offset_2, dims)::AbstractArray

Creates a view of weights from offset_1 to offset_2 with the dimension specified by dims.

Llama2Inference.memory_map_weights — Method

memory_map_weights(config::Config, weights::Vector{T}, shared_weights::Int32)::TransformerWeights{T} where {T<:AbstractFloat}

Takes the values in weights and maps creates a new TransformerWeights struct using config and shared_weights.

Llama2Inference.RunState — Method

RunState(config::Config)

Create a RunState object with the given configuration.

This is used to store the state of the model during inference and is 0-initialized.